requirements.txt
Flask==2.0.1
Flask-MySQLdb==2.0.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.26.0
Pillow==8.4.0
matplotlib==3.4.3
uuid==1.30
Flask-WTF==1.0.0
Flask-Limiter==2.4.0
email-validator==1.1.3